+++ CHANGES.md
@@ -140,3 +140,4 @@ some files renamed or trashed. In addtion, the "missing" files are still kept ba
 backups even if not linked anymore).
 * fix: deletion of a directory contents now deletes the hidden files too.
 * fix (0.97.1): then link to the "Camera Club de Genève" s'ouvre dans une fenêtre dédiée.
+* fix (0.97.2): fixes possible issue when a sort is done during refresh of a library.